tool/requirements.txt
attrs==25.1.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
idna==3.10
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
lazy-object-proxy==1.10.0
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
packaging==24.2
pathable==0.4.4
prance==23.6.21.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986==2.0.0
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
six==1.17.0
timeout-decorator==0.5.0
typing_extensions==4.12.2
urllib3==2.3.0
